refactor(natives): modularize subagent control domains
